requirements.txt
streamlit==1.28.2
psycopg2-binary==2.9.5
sqlalchemy==2.0.27
requests==2.31.0
python-dotenv==1.0.0
# psycopg2==2.9.9
# scikit-learn==0.22.2.post1
# protobuf==3.20.*
# click==8.0.4
